python-server/pyproject.toml
[project]
name = "handinhand-python-server"
version = "1.0.0"
description = "Hand in Hand Game Server - Python FastAPI Implementation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "redis>=5.0.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
serve = "handinhand_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]